+++ src/domo/policy.rs
@@ -2,14 +2,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Policy {
-    pub id: u32,
+    pub id: usize,
     pub name: String,
     #[serde(rename = "type")]
     pub policy_type: String,
     #[serde(rename = "users")]
-    pub user_ids: Vec<u64>,
+    pub user_ids: Vec<usize>,
     #[serde(rename = "groups")]
-    pub group_ids: Vec<u64>,
+    pub group_ids: Vec<usize>,
     pub filters: Vec<Filter>,
 }
